this version contains ratings controllers and routes, I have added one model, one route, and the function responsable of the rating process is in the UserController

+++ database/migrations/2022_12_08_151148_create_experts_table.php
@@ -16,10 +16,10 @@ return new class extends Migration
         Schema::create('expert_detail', function (Blueprint $table) {
             $table->id();
             $table->string('skills');
-            $table->integer('rating')->default(0);
+            $table->double('rating')->default(0);
             $table->integer('ratings')->default(0);
             $table->double('cost');
-            
+
             $table->unsignedBigInteger('user_id')->uniqid();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
